+++ Modules/CodeTrek/Resources/views/layouts/master.blade.php
@@ -5,4 +5,4 @@
 @section('css_scripts')
 <link href="{{ mix('css/codetrek.css') }}" rel="stylesheet">
 @endsection
-<input type="hidden" id="get_report_data_url" value={{ route('codetrek.get-report-data') }}>
+<input type="hidden" id="get_report_data_url" value={{ route('reports.codetrek.get-report-data') }}>
