+++ Modules/CodeTrek/Routes/web.php
@@ -13,7 +13,6 @@
 
 Route::prefix('codetrek')->middleware('auth')->group(function () {
     Route::get('/', 'CodeTrekController@index')->name('codetrek.index');
-    Route::get('get-report-data', 'CodeTrekController@getApplicantData')->name('codetrek.get-report-data');
     Route::post('/', 'CodeTrekController@store')->name('codetrek.store');
     Route::post('/store-feedback', 'CodeTrekController@storeCodeTrekApplicantFeedback')->name('codetrek.store-feedback');
     Route::get('/edit/{applicant}', 'CodeTrekController@edit')->name('codetrek.edit');
