+++ Modules/Report/Routes/web.php
@@ -19,6 +19,7 @@ Route::prefix('report')->group(function () {
     Route::get('/delete/{id}', 'ReportController@delete')->name('report.delete');
     Route::post('/update/{id}', 'ReportController@update')->name('report.update');
     Route::get('/get-fte-report/{user}', 'UserReportController@getFteData')->name('reports.fte.get-report-data');
+    Route::get('get-report-data', 'ReportController@getApplicantData')->name('reports.codetrek.get-report-data');
 
     Route::prefix('finance')->group(function () {
         Route::get('dashboard', 'FinanceReportController@dashboard')->name('reports.finance.dashboard');
