+++ Modules/Report/Services/Finance/ReportDataService.php
@@ -3,6 +3,7 @@
 namespace Modules\Report\Services\Finance;
 
 use Modules\Client\Entities\Client;
+use Modules\CodeTrek\Entities\CodeTrekApplicant;
 
 class ReportDataService
 {
@@ -22,6 +23,31 @@ class ReportDataService
         }
 
         return $filters;
+    }
+
+    public function getDataForCodeTrekApplications($type, $filters, $request)
+    {
+        if ($type == 'codetrek-application') {
+            $applicants = CodeTrekApplicant::find('start_date');
+            $defaultStartDate = $applicants->created_at ?? today()->subYear();
+            $defaultEndDate = today();
+            $filters['start_date'] = empty($filters['start_date']) ? $defaultStartDate : $filters['start_date'];
+            $filters['end_date'] = empty($filters['end_date']) ? $defaultEndDate : $filters['end_date'];
+            $applicantChartData = CodeTrekApplicant::select(\DB::Raw('DATE(start_date) as date, COUNT(*) as count'))
+                ->whereDate('start_date', '>=', $filters['start_date'])
+                ->whereDate('start_date', '<=', $filters['end_date'])
+                ->groupBy('date');
+
+            $dates = $applicantChartData->pluck('date')->toArray();
+            $counts = $applicantChartData->pluck('count')->toArray();
+            $chartData = [
+                'dates' => $dates,
+                'counts' => $counts,
+            ];
+            $reportApplicantData = json_encode($chartData);
+
+            return $reportApplicantData;
+        }
     }
 
     public function getDataForClientRevenueReportPage(array $data)
